Interpolation suite complete
Interpolation suite has been totally flushed out. Each method is testable (with the exception of WENO, probably removing as interpolator and making a constructor instead.)
Wide variety of test functions and L-2 and L-\infty norms are computed, absolute and rms error are computed as well, to instill confidence in performance.